add compiler optimization level 3, some improvements and a performance test

+++ Sources/ObjC/Vector.mm
@@ -1,7 +1,7 @@
 #import "Vector.h"
 
 @implementation Vector
-- (instancetype) initWithX: (double) x y:(double) y {
+- (instancetype) initWithX: (float) x y:(float) y {
     self = [super init];
     self.x = x;
     self.y = y;
