+++ Sources/ObjC/include/Vector.h
@@ -3,9 +3,9 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @interface Vector : NSObject
-@property (nonatomic) double x;
-@property (nonatomic) double y;
-- (instancetype) initWithX: (double) x y:(double) y;
+@property (nonatomic) float x;
+@property (nonatomic) float y;
+- (instancetype) initWithX: (float) x y:(float) y;
 @end
 
 NS_ASSUME_NONNULL_END
